separate API types to new type package (for reuse in middlewares and tests)
