+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
     "go.buildFlags": [
-        "-tags=unit",
+        "-tags=integration",
         "-v"
     ],
-    "go.testTags": "unit",
+    "go.testTags": "integration",
 }
